Backend de Educacion

+++ backend/src/main/java/com/argentinaprogramaanrd/anrd/Controller/PersonaController.java
@@ -5,7 +5,6 @@
 package com.argentinaprogramaanrd.anrd.Controller;
 
 import com.argentinaprogramaanrd.anrd.Entity.Persona;
-import com.argentinaprogramaanrd.anrd.Interface.IPersonaService;
 import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.access.annotation.Secured;
@@ -17,6 +16,11 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
+import com.argentinaprogramaanrd.anrd.Security.Controller.Helpers.Message;
+import com.argentinaprogramaanrd.anrd.Service.PersonaService;
+import org.apache.commons.lang3.StringUtils;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 
 /**
  *
@@ -27,34 +31,75 @@ import org.springframework.web.bind.annotation.RestController;
 public class PersonaController {
 
     @Autowired
-    IPersonaService iPersonaService;
+    PersonaService iPersonaService;
 
     @GetMapping("personas")
-    public List<Persona> index() {
-        return iPersonaService.index();
+    public ResponseEntity<List<Persona>> index() {
+        List<Persona> list = iPersonaService.index();
+        return new ResponseEntity(list, HttpStatus.OK);
     }
 
     @GetMapping("personas/show/{id}")
-    public Persona show(@PathVariable Long id) {
-        return iPersonaService.show(id);
+    public ResponseEntity<Persona> show(@PathVariable("id") int id) {
+        if (!iPersonaService.existsById(id)) {
+            return new ResponseEntity(new Message("La persona que buscas no existe."), HttpStatus.BAD_REQUEST);
+        }
+        return new ResponseEntity(iPersonaService.show(id).get(), HttpStatus.OK);
+
     }
 
     //@PreAuthorize("hasRole('ADMIN')")
     @Secured("ROLE_ADMIN")
     @PostMapping("personas")
-    public String save(@RequestBody Persona persona) {
+    public ResponseEntity<?> store(@RequestBody Persona persona) {
+
+        if (StringUtils.isBlank(persona.getName())) {
+            return new ResponseEntity(new Message("El nombre es obligatorio"), HttpStatus.BAD_REQUEST);
+        }
+
+        if (StringUtils.isBlank(persona.getLast_name())) {
+            return new ResponseEntity(new Message("El apellido es obligatorio"), HttpStatus.BAD_REQUEST);
+        }
+
+        if (StringUtils.isBlank(persona.getJob())) {
+            return new ResponseEntity(new Message("El trabajo es obligatorio"), HttpStatus.BAD_REQUEST);
+        }
+
+        if (StringUtils.isBlank(persona.getDescription())) {
+            return new ResponseEntity(new Message("La descripción es obligatoria"), HttpStatus.BAD_REQUEST);
+        }
+
         iPersonaService.save(persona);
-        return "Persona creada correctamente";
+        return new ResponseEntity(new Message("Persona Agregada Correctamente."), HttpStatus.OK);
+
     }
 
     //@PreAuthorize("hasRole('ADMIN')")
     @Secured("ROLE_ADMIN")
     @PutMapping("personas/update/{id}")
-    public Persona edit(@PathVariable Long id,
-            @RequestBody Persona persona
-    ) {
-        Persona old_persona = iPersonaService.show(id);
+    public ResponseEntity<?> update(@PathVariable int id, @RequestBody Persona persona) {
 
+        if (!iPersonaService.existsById(id)) {
+            return new ResponseEntity(new Message("La persona que buscas no existe."), HttpStatus.BAD_REQUEST);
+        }
+
+        if (StringUtils.isBlank(persona.getName())) {
+            return new ResponseEntity(new Message("El nombre es obligatorio"), HttpStatus.BAD_REQUEST);
+        }
+
+        if (StringUtils.isBlank(persona.getLast_name())) {
+            return new ResponseEntity(new Message("El apellido es obligatorio"), HttpStatus.BAD_REQUEST);
+        }
+
+        if (StringUtils.isBlank(persona.getJob())) {
+            return new ResponseEntity(new Message("El trabajo es obligatorio"), HttpStatus.BAD_REQUEST);
+        }
+
+        if (StringUtils.isBlank(persona.getDescription())) {
+            return new ResponseEntity(new Message("La descripción es obligatoria"), HttpStatus.BAD_REQUEST);
+        }
+
+        Persona old_persona = iPersonaService.show(id).get();
         old_persona.setName(persona.getName());
         old_persona.setLast_name(persona.getLast_name());
         old_persona.setJob(persona.getJob());
@@ -62,14 +107,20 @@ public class PersonaController {
         old_persona.setImage(persona.getImage());
 
         iPersonaService.save(old_persona);
-        return old_persona;
+        return new ResponseEntity(new Message("Persona creada correctamente."), HttpStatus.OK);
+
     }
 
     @Secured("ROLE_ADMIN")
     @DeleteMapping("personas/delete/{id}")
-    public String delete(@PathVariable Long id) {
+    public ResponseEntity<?> delete(@PathVariable int id) {
+
+        if (!iPersonaService.existsById(id)) {
+            return new ResponseEntity(new Message("El id de la experiencia no existe."), HttpStatus.BAD_REQUEST);
+        }
+
         iPersonaService.delete(id);
-        return "Persona eliminada correctamente";
+        return new ResponseEntity(new Message("Persona se eliminó correctamente."), HttpStatus.OK);
 
     }
 }
